Squash-to: "mptcp: Implement path manager interface  commands"
Implement stubs for PM events delegating the action to the work
queue. This allows acquiring whatever lock is needed to perform
the actual implementation.

Try to avoid scheduling the worker if no action is needed/possible.
I relies on the accounting info included into the PM struct and
on great deal of double-checked locking [anti-]pattern.

v1  -> v2:
 - fix several issues in mptcp_pm_addr_signal()

RFC -> v1:
 - simplify/cleanup mptcp_pm_work_pending() - Mat
 - likewise simplify/cleanup mptcp_pm_add_addr()

Signed-off-by: Paolo Abeni <pabeni@redhat.com>