mptcp: add netlink based PM
Expose to U/S a netlink family to control the PM, setting:

 - list of local addresses to be signaled.
 - list of local addresses used to created subflows.
 - maximum number of add_addr option to react

When the msk is fully established, the PM netlink attempts to
create subflow for each addr in 'local' list, waiting for each
connection to be completed before attempting the next one.

After exausting the 'local' list, the PM tries to announce the
'signal' list via the ADD_ADDR option. Since we currenlty lack
the ADD_ADDR echo (and related event) only the first addr is sent.

Idea is to add an additional PM hook for ADD_ADDR echo, to allow
the PM netlink announcing multiple addresses, in sequence.

v1  -> v2:
 - explicitly reset pm status to idle after processing a command
 - fix msk pm initialization
 - fix mptcp_pm_nl_add_addr bugs
 - added ifindex support
 - hopefully clearer local function names
 - always use  IS_ENABLED(CONFIG_MPTCP_IPV6) instead of
    IS_ENABLED(CONFIG_IPV6)

RFC -> v1:
 - simplified NL API
 - reduced {WRITE,READ}_ONCE boilerplate due to PM changes
 - add check for duplicate addresses

Signed-off-by: Paolo Abeni <pabeni@redhat.com>