mptcp: allow picking different xmit subflows
Update the scheduler to less trivial heuristic: cache
the last used subflow, and try to send on it a reasonably
long burst of data. When the burst or the subflow send
space is exausted, move to the next one with available
send space.

Signed-off-by: Paolo Abeni <pabeni@redhat.com>
---
RFC -> v1:
 - simplify mptcp_subflow_active() - Florian
 - use sk_stream_memory_free() to check for free space

Note: still keep the sk_state == TCP_CLOSE chk in __mptcp_move_skbs():
the alternative would be clearing data_avail at close time.
That should happen after closing the subflow, but for outgoing
connection, the subflow socket could be already free -
we call sock_release().