Squash-to: "mptcp: allow picking different xmit subflows"
This improves the current packet scheduler for scenarios
where the msk sockets is using different links with different
bandwith limits.

The commit comment should be changed into:

"""
Update the scheduler to less trivial heuristic: cache
the last used subflow, and try to send on it a reasonably
long burst of data.

When the burst or the subflow send space is exhausted, pick
the subflow with the lower ratio between write space and
send buffer - that is, the subflow with the greater relative
amount of free space.
"""

Signed-off-by: Paolo Abeni <pabeni@redhat.com>
---
Note:
- this is almost an RFC ;)
- subflows are now selected when sk_stream_memory_free(ssk),
  while before we used sk_stream_is_writeable(ssk). The current
  choice is closer to plain TCP and IMHO more correct, but
  even the old one should not be buggy.
  Anyhow I saw reproducible write error (EAGAIN) when using
  sk_stream_is_writeable() with this strategy, and I was unable
  to track them down, so possibly there is some other bug there
  I could not find.
- simuilt xmit self-tests are still failing, even if we are now
  closer to success. The failures are now caused by the high delay
  configured on the virtual link.
  This problem would be addressed by a larger initial send buffer,
  but the current xmit path is unable to cope with that correctly:
  with a larger send buffer it will send pkts outside the MPTCP
  window.
  To properly address the above we additionally need something alike
  "mptcp: keep track of receivers advertised window" from Florian,
  plus the ability to enqueue data at the MPTCP level and send it
  when the MPTCP level window is open.