mptcp: use mptcp backlog.
We have some tasks triggered by the subflow receive path
which require to access the msk socket status, specifically:
mptcp_clean_una() and mptcp_push_pending()

We have almost everything in place to use the msk backlog
and a standard !owned by user schema for that.

Introduce a dummy per msk sk_buff to trigger into the backlog
as needed.

Since the worker is no more used to clean the acked data,
for fallback sockets we need to explicitly flush them.

As added bonus we can simplify a lot mptcp_poll() and
move the timer update under the data lock.

The worker is now used only to process and send DATA_FIN
packets and do the mptcp-level retransmissions.

Signed-off-by: Paolo Abeni <pabeni@redhat.com>
--
v2 -> v3:
 - move clean una to release_cb(), so that we don't need
   an additional spin lock contention

v1 -> v2
 - add tx_pending_data accounting in __mptcp_subflow_push_pending()
 - in mptcp_data_acked schedule explicitly schedule the worker if
   ack is pending - checking sk status is both too liberal and not
   enough.
 - add missing call to __mptcp_check_send_data_fin() in mptcp_worker()
   this is needed to cope with __mptcp_subflow_push_pending() deferring
   send ack to the worker
 - renamed backlog to bl_entry
 - sk_mem_reclaim_partial() -> __mptcp_mem_reclaim_partial() in
   mptcp_clean_una(). We are both under the data_lock and the socket
   lock so we can call the more effective mptcp variant
 - do not try to push pending frames on wnd update if no pending data
   (use the new __mptcp_wnd_updated helper for that)
 - move write queue cleanup in destroy_common() so that it takes place
   after that subflows stops touching msk->sk_forward_alloc.
   and we don't need to acquire the msk socket lock spin lock to reclaim
   the xmit memory
 - explicitly init info->flags in __mptcp_subflow_push_pending(), or
   horrible corruption will happen later
 - explicitly flush pending data for fallback socket on ack reception.
   This likely fixes an older issue, which was hidden by the worker,
   spooling the pending data, just slower.